[yaml] refactoring: rename intellij.yaml to intellij.yaml.backend (IJPL-176102)
This module isn't used in a frontend variant of the IDE, so it should use the '.backend' suffix for consistency and to avoid confusion.

References from the build script to the 'intellij.yaml' module aren't changed in this commit because 'intellij.yaml.editing' module will be renamed to 'intellij.yaml' in the next commit, and plugin.xml will be moved to it.

GitOrigin-RevId: 1c535a9a02bdccd89f46cdad71e34064b1dab854